pyproject.toml
[tool.poetry]
name = "Evolution"
version = "0.1.0"
description = "Simulating rabbits evolution"
authors = ["Milan Gittler <gittlermilan@seznam.cz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">= 3.10, <3.11"
simpy = "^4.0.2"
matplotlib = "^3.8.1"
pydantic = "^2.5.2"
pillow = "^10.1.0"
pandas = "^2.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"